Move token expiry date validation to TokenService
