+++ src/main/resources/messages_en.properties
@@ -9,7 +9,7 @@ skill.notfound=Skill does not exist
 token.notfound=Token does not exist
 user.notfound=User does not exist
 userCv.notfound=User cv does not exist
-
+token.expired=Link expired
 mail.confirmation.title= Confirmation email
 mail.confirmation.message= Click to below link to confirm you account
 mail.confirmation.link= http://localhost:8080/api/token?value=
